[clang][NFC] Refactor `Selector` to use `PointerIntPair` inside
